% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_sales_df}
\alias{get_sales_df}
\alias{get_org_names}
\title{Get a sales data frame for demo}
\usage{
get_sales_df()

get_org_names()
}
\description{
Get a sales data frame for demo
}
\references{
\href{https://www.kaggle.com/datasets/surajjha101/stores-area-and-sales-data?resource=download}{Supermarket store branches sales analysis data set}
}
