% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_filters.R
\name{select_org}
\alias{select_org}
\title{Get a sales data frame for demo}
\usage{
select_org(label, selectize, multiple, selected, width, size)
}
\arguments{
\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{selectize}{Whether to use \pkg{selectize.js} or not.}

\item{multiple}{Is selection of multiple items allowed?}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}). If not specified then defaults to the first value for
single-select lists and no values for multiple select lists.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:reexports]{validateCssUnit()}}.}

\item{size}{Number of items to show in the selection box; a larger number
will result in a taller box. Not compatible with \code{selectize=TRUE}.
Normally, when \code{multiple=FALSE}, a select input will be a drop-down list,
but when \code{size} is set, it will be a box instead.}
}
\description{
Get a sales data frame for demo
}
